# PostingsDecodingUtil: interchange loops to enable better memory access and SIMD vectorisation