--- a/lucene/core/src/java/org/apache/lucene/internal/vectorization/PostingDecodingUtil.java
+++ b/lucene/core/src/java/org/apache/lucene/internal/vectorization/PostingDecodingUtil.java
@@ -44,14 +44,21 @@ public class PostingDecodingUtil {
   public void splitInts(
       int count, int[] b, int bShift, int dec, int bMask, int[] c, int cIndex, int cMask)
       throws IOException {
-    // Default implementation, which takes advantage of the C2 compiler's loop unrolling and
-    // auto-vectorization.
     in.readInts(c, cIndex, count);
-    int maxIter = (bShift - 1) / dec;
-    for (int i = 0; i < count; ++i) {
-      for (int j = 0; j <= maxIter; ++j) {
-        b[count * j + i] = (c[cIndex + i] >>> (bShift - j * dec)) & bMask;
+    final int maxIter = (bShift - 1) / dec;
+
+    // Process each shift level across all elements (better for vectorization)
+    for (int j = 0; j <= maxIter; ++j) {
+      final int shift = bShift - j * dec;
+      final int bOffset = count * j;
+      // Vectorizable loop: contiguous memory access with simple operations
+      for (int i = 0; i < count; ++i) {
+        b[bOffset + i] = (c[cIndex + i] >>> shift) & bMask;
       }
+    }
+
+    // Apply mask to c array (vectorizable)
+    for (int i = 0; i < count; ++i) {
       c[cIndex + i] &= cMask;
     }
   }