@jpountz 2025-08-22
Should we do the same in `MemorySegmentPostingDecodingUtil`?